requirements.txt
python-jobspy
pandas
requests
beautifulsoup4
termcolor
tabulate
